Re-integrate all parts of Allocator into one file
